While and For Loops
Added the actions for while and for loops. In order to facilitate the passing of labels, I also added to tree.h a structure that holds the labels for the conditional sections and for the end of the loop, along with an enumeration that stores the type of for loop (either for...to or for...downto).